chore(db): archive agent run events table
