+++ turbo/packages/db/src/index.ts
@@ -3,7 +3,7 @@ import * as deviceCodesSchema from "./schema/device-codes";
 import * as cliTokensSchema from "./schema/cli-tokens";
 import * as agentComposeSchema from "./schema/agent-compose";
 import * as agentRunSchema from "./schema/agent-run";
-import * as agentRunEventSchema from "./schema/agent-run-event";
+import * as archivedAgentRunEventSchema from "./schema/archived-agent-run-event";
 import * as conversationSchema from "./schema/conversation";
 import * as checkpointSchema from "./schema/checkpoint";
 import * as agentSessionSchema from "./schema/agent-session";
@@ -81,7 +81,7 @@ export const schema = {
   ...cliTokensSchema,
   ...agentComposeSchema,
   ...agentRunSchema,
-  ...agentRunEventSchema,
+  ...archivedAgentRunEventSchema,
   ...conversationSchema,
   ...checkpointSchema,
   ...agentSessionSchema,
